+++ classes/SortedFolder.py
@@ -1,0 +1,4 @@
+class SortedFolder:
+    def __init__(self, name, associated_types):
+        self.name = name
+        self.associated_types = associated_types
